# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shopee Affiliate Platform API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface FaqItem {
  question: string;
  answer: string;
}

export interface Product {
  id: string;
  shopeeId: string;
  name: string;
  slug: string;
  price: number;
  priceBeforeDisc?: number | null;
  commission?: number | null;
  commissionRate?: number | null;
  imageUrl: string;
  images?: string[] | null;
  shopName?: string | null;
  shopRating?: number | null;
  ratingStar?: number | null;
  soldCount?: number | null;
  category?: string | null;
  tags?: string[] | null;
  affiliateLink: string;
  reviewContent?: string | null;
  pros?: string[] | null;
  cons?: string[] | null;
  faq?: FaqItem[] | null;
  metaTitle?: string | null;
  metaDesc?: string | null;
  viewCount: number;
  clickCount: number;
  conversionCount: number;
  status: string;
  createdAt: string;
  lastUpdated: string;
  publishedAt?: string | null;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface GenerateAffiliateBody {
  url: string;
}

export interface AffiliateGenerateResponse {
  product: Product;
  isNew: boolean;
}

export interface ClickTrackRequest {
  referer?: string | null;
}

export interface ClickTrackResponse {
  success: boolean;
  affiliateLink: string;
}

export interface UpdateProductBody {
  reviewContent?: string | null;
  pros?: string[] | null;
  cons?: string[] | null;
  faq?: FaqItem[] | null;
  metaTitle?: string | null;
  metaDesc?: string | null;
  tags?: string[] | null;
  status?: string | null;
}

export interface GenerateContentBody {
  productId?: string | null;
  productName: string;
  productCategory?: string | null;
  productDescription?: string | null;
  price?: number | null;
}

export interface GenerateContentResponse {
  reviewContent: string;
  pros: string[];
  cons: string[];
  faq: FaqItem[];
  metaTitle: string;
  metaDesc: string;
  tags: string[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardStats {
  totalProducts: number;
  publishedProducts: number;
  draftProducts: number;
  totalClicks: number;
  totalViews: number;
  totalConversions: number;
  estimatedCommission: number;
  topCategories: CategoryCount[];
  recentProducts: Product[];
}

export interface TopProduct {
  id: string;
  name: string;
  slug: string;
  imageUrl: string;
  clickCount: number;
  viewCount: number;
  commission?: number | null;
  commissionRate?: number | null;
  price: number;
  status: string;
}

export interface ClickChartPoint {
  date: string;
  clicks: number;
  views: number;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
  message: string;
}

export interface Setting {
  id: string;
  key: string;
  value?: string | null;
  description?: string | null;
  updatedAt: string;
}

export interface UpsertSettingBody {
  key: string;
  value: string;
  description?: string | null;
}

export type ListProductsParams = {
  page?: number;
  limit?: number;
  category?: string;
  sort?: ListProductsSort;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  newest: "newest",
  cheapest: "cheapest",
  rating: "rating",
  popular: "popular",
} as const;

export type GetTopProductsParams = {
  limit?: number;
};

export type GetClicksChartParams = {
  days?: number;
};

export type SearchProductsParams = {
  q?: string;
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  sort?: SearchProductsSort;
  page?: number;
  limit?: number;
};

export type SearchProductsSort =
  (typeof SearchProductsSort)[keyof typeof SearchProductsSort];

export const SearchProductsSort = {
  newest: "newest",
  cheapest: "cheapest",
  rating: "rating",
  popular: "popular",
} as const;

export type AdminListProductsParams = {
  page?: number;
  limit?: number;
  status?: AdminListProductsStatus;
};

export type AdminListProductsStatus =
  (typeof AdminListProductsStatus)[keyof typeof AdminListProductsStatus];

export const AdminListProductsStatus = {
  draft: "draft",
  published: "published",
  all: "all",
} as const;